Rename file from valorant_adapter to vlr_adapter
